Struct lrucache::arena::Arena

pub struct Arena<T> {
    items: Vec<Entry<T>>,
    capacity: usize,
    generation: u64,
    free_list_head: Option<usize>,
}

A generational arena for allocating memory based off a vector. Every entry is associated with a generation counter to uniquely identify newer allocations from older reclaimed allocations at the same position in the vector. This is inspired from the crate “generational-arena”

Fields
items: Vec<Entry<T>>capacity: usizegeneration: u64free_list_head: Option<usize>
Implementations

impl<T> Arena<T>

pub fn new() -> Self

pub fn capacity(&self) -> usize

pub fn reserve(&mut self, capacity: usize)

pub fn with_capacity(capacity: usize) -> Self

pub fn insert(&mut self, item: T) -> Result<Index, ArenaOOM>

pub fn remove(&mut self, index: &Index) -> Option<T>

pub fn get_mut(&mut self, index: &Index) -> Option<&mut T>

pub fn get(&self, index: &Index) -> Option<&T>